feat: 이름변경 및 참가자 API 생성
