+++ src/types/parties.types.ts
@@ -5,6 +5,13 @@ export type Recruit = Tables<'party_recruit'>;
 
 export type RecruitForm = Pick<Recruit, 'party_type' | 'title' | 'description'>;
 
+export type PartyMember = Tables<'party_member'>;
+
 export interface RecruitWithProfile extends Omit<Recruit, 'created_by'> {
   created_by: Profile;
+}
+
+export interface PartyMemberWithProfile
+  extends Omit<PartyMember, 'profile_id'> {
+  profile_id: Profile;
 }
